Add open source firmware
This is a new firmware that was rewritten from scratch in Verilog.
The free Quartus Prime Lite 20.1 was used to create and build it.

The firmware has only been test briefly yet. Some features are missing,
such as support for bus arbitration, function codes, and there is no
support for resetting the firmware state machines after the CPLD has
been powered on.

The protocol between Raspberry Pi and CPLD will undergo changes,
so this version of the firmware is a base for building on.